+++ packages/__tests__/src/legacy/scrollTime.ts
@@ -36,4 +36,20 @@ describe('scrollTime', () => {
     expect(scrollTop).toBeGreaterThan(0)
     expect(diff).toBeLessThan(3)
   })
+
+  it('doesn\'t get applied on navigation when preserveScroll is false', () => {
+    let calendar = initCalendar({
+      scrollTime: '02:00:00',
+      preserveScroll: false,
+      height: 400, // short enough to make scrolling happen
+    })
+    let viewWrapper = new TimeGridViewWrapper(calendar)
+    let scrollEl = viewWrapper.getScrollerEl()
+
+    scrollEl.scrollTop = 99999
+    let scrollTop = scrollEl.scrollTop
+
+    calendar.next()
+    expect(scrollEl.scrollTop).toBe(scrollTop) // stays the same
+  })
 })